refactor: coupon context로직 분리
